gendash-engine/llm.py
from langchain_google_genai import ChatGoogleGenerativeAI
import os
from dotenv import load_dotenv
load_dotenv()
llm = ChatGoogleGenerativeAI(api_key=os.getenv("GOOGLE_API_KEY"), model="gemini-flash-latest")
response = llm.invoke("Sing a ballad of LangChain.")
print(response.content)