gendash-engine/requirements.txt
# Core Framework
flask==3.0.0
flask-cors==4.0.0

# LangChain and AI
langchain==0.1.0
langchain-google-genai==0.0.6
google-generativeai==0.3.2

# Data Processing
pandas==2.1.4
numpy==1.26.2

# Validation and Schema
pydantic==2.5.3

# HTTP Requests
requests==2.31.0

# Retry Logic
tenacity==8.2.3

# Environment Variables
python-dotenv==1.0.0

# Production Server
gunicorn==21.2.0
